improve contentEditable for IE 8
I don't want to spend more time on IE 8.
So it would be the ending fixup. And i will still leave related code for
a while maybe in case anyone want to help to improve it.
Just encourge your users to upgrate the browers or just switch to a
batter one please !!