fix #252 default matcher will match the `'` now
for IME.
IME will insert a `'` char to separate the word, it will make the
match fail. So the easy way is adding it to matcher.

There should be a `isComposing` property of keyboardEvent in additon to
W3C standard when the event is between the `compostionstart` and the
`compositionend`, but only firefox support it.
so let's give it an easy fix for now, and i think it's good enough.