Do not hide view on scroll events if scrollTop hasn't been changed
There is a bug in Firefox that causes the suggestion menu to hide
immediately after shown when when the parent of the textarea has
the display: table style. Caret.js triggers the bug by adding an
element to the parent of the textarea. When this happens, Firefox
fires a scroll event on the textarea. Luckily, when this bug happens
the scrollTop value remains the same.

This commit records the scrollTop value everytime a scroll event is
dispatched, and compare it to the last value to ensure it has been
changed before hiding the view.

See also:
Firefox bug tracker entry:
  https://bugzilla.mozilla.org/show_bug.cgi?id=1175351