Change zwj to zws
In `at.js` the zero width joiner unicode character (\u200D) is inserted after an
at-mention is inserted into the rich text editor (for our use case, the
at-mention is a condition placeholder). This zwj character causes problems with
the cursor - additionally, the use case for the zwj character is very specific
and in its current form does not do anything. By switching to a zero width space
character we can preserve the spirit of the original implementation and fix the
cursor issue.

This was tested on Google Chrome, Firefox, and Safari. The zwj character was
added to `at.js` in this PR: https://github.com/zendesk/At.js/pull/7 - it's
important to make sure we are not introducing any regressions with this change.